requirements.txt
flask
flask_sqlalchemy
werkzeug
google-generativeai
markdown
pillow
